Synchronously flush Gui upon SetManager
This enables clients to assume that after SetManager is called, the
manager has changed and that the Layout function (which sets the views)
has been called. AFIK without this behavior, there are no known
guarantees that later code dependent on the change in manager will work,
partly because the resize event previously sent is not handled* and
partly because Update()'s semantics do not offer total order.

*Perhaps an alternative to this commit is to actually handle this resize
event, but blocking until Layout has been called is the easiest to code
against (otherwise all other modifications to the Gui would have to go
through Update(), even if no concurrency was used. Plus Update() calls
are not guaranteed to be totally ordered so there I believe there is no
race-free way to use SetManager because of this :P).