Change semantics of Gui.Update() to include strict total order
Previously, Gui did not guarantee that a Update() occurring before
another Update() would cause the first to execute before the second
(i.e. offering a strict total order with respect to the calling order
of Update()s). This was due to Update() offering non-blocking
semantics and the use of a go routine to prevent sends on a channel
from blocking (since channels are buffered and block when full).

This makes the API very unpleasant to use in a concurrent context.
Workarounds such as nesting Update() calls work but hurt readability and
are unintuitive IMO.

This commit introduces the natural and much desired semantics of strict
total ordering when using Update() while keeping the non-blocking nature
of Update(). This is achieved using a condition variable based on the
calling order of each Update() within each go routine. Admittedly, this
is achieved using a somewhat heavy handed approach but it should be
sufficient enough for this module's purposes.

Pretty Neat! :^)